Create separate app_component.ts
